fix: felt representation mismatch between rust and miden
In the Miden VM, field elements are the machine word type, which as a
concrete value, is typically stored as a 64-bit integral value. However
the VM is best treated as a 32-bit machine when viewed from the lens of
a typical compiler. This is because its pointer type must be a valid
32-bit integer, many of the native operations are only available on the
32-bit integral type (most notably, bitwise operators, and two's
complement arithmetic).

So when determining how to represent Miden to Rust, we chose to do so
as a 32-bit machine. A major motivation for this was to be able to
compile from Rust to Miden via Wasm, which is itself a 32-bit machine
(there is a 64-bit variant, but it is not ready for primetime yet).

The problem we run into when interacting with Miden from Rust, is with
the field element type (felt). Because felts ostensibly require 64 bits
to fully represent, but are also used as the machine word type on Miden,
we have a catch-22: we can't faithfully represent felt values in Rust,
while also presenting Miden as a 32-bit machine. Additionally, we are
constrained in the need to be able to identify values in Wasm, which are
being used as felts, during compilation.

More concretely: if we represent felts in Rust as 64-bit values, then we
cannot pass them as-is to Miden procedures, because a 64-bit value in
Rust's memory model gets translated to _two_ field elements when
lowering to Miden Assembly (in order to emulate byte-addressable memory
in Miden, we rely on the u32 instruction set, which means we are forced
to treat every memory cell, a felt, as a 32-bit chunk of memory). This
also would necessarily affect aggregate types, such as words.

The solution then, at least as envisioned here, is to make felts an
opaque type of 32-bits in Rust, and use a set of compiler intrinsics to
allow constructing felt values from u64 (or vice versa). Additionally,
we use the f32 type to signal that a value on the Wasm value stack is
actually a felt, not a regular integral type. Because of how we lower
Rust code to Miden, this actually works - the "extra" information stored
in each field element is effectively hidden from Rust, but remains
accessible from Miden Assembly. This is because Rust thinks they are
only 32-bit values, but the actual machine word size allows the full
field to be represented. In essence, we are hiding the real felt value
in the gap between how Rust thinks Miden represents things, and how
Miden actually represents things.

+++ hir-type/src/layout.rs
@@ -5,8 +5,8 @@ use smallvec::SmallVec;
 
 use super::*;
 
-const FELT_SIZE: usize = core::mem::size_of::<u64>();
-const WORD_SIZE: usize = core::mem::size_of::<[u64; 4]>();
+const FELT_SIZE: usize = core::mem::size_of::<u32>();
+const WORD_SIZE: usize = core::mem::size_of::<[u32; 4]>();
 
 impl Type {
     /// Convert this type into a vector of types corresponding to how this type
